Extract remove, ban, unban into moderation
